Refactored FoodItemCards
