--- order-system-ionic/src/components/FoodItemCard.css
@@ -1,29 +1,0 @@
-/* ion-card {
-    border-radius: 1rem;
-    box-shadow: none;
-    display: flex;
-    flex-direction: row;
-}
-
-ion-card-header {
-    padding-bottom: 0;
-}
-
-ion-card-title {
-    font-size: 1.5rem;
-    font-weight: bold;
-}
-
-ion-card-subtitle {
-    font-size: 1rem;
-}
-
-ion-icon {
-    font-size: 1rem;
-    margin-right: 0.1rem;
-}
-
-ion-button {
-    --shape: round;
-    --border-radius: 2.0rem;
-} */
