+++ order-system-ionic/src/components/FoodItemCards.tsx
@@ -1,0 +1,6 @@
+import CartFoodItemCard from "./FoodItemCards/CartFoodItemCard";
+import MenuFoodItemCard from "./FoodItemCards/MenuFoodItemCard";
+import OrderFoodItemCard from "./FoodItemCards/OrderFoodItemCard";
+import SplitBillFoodItemCard from "./FoodItemCards/SplitBillFoodItemCard";
+
+export { MenuFoodItemCard, OrderFoodItemCard, CartFoodItemCard, SplitBillFoodItemCard };
