+++ order-system-ionic/src/components/FoodItemCards/DietIcons.tsx
@@ -19,4 +19,3 @@ const DietIcons: React.FC<DietProps> = ({ vegan, vegetarian, glutenFree }) => {
 
 export default DietIcons;
 export type { DietProps };
-export { DietIcons };
