+++ order-system-ionic/src/components/FoodItemCards/Utils.tsx
@@ -1,0 +1,6 @@
+const formatPrice = (price: number) => {
+    const priceStr = price.toFixed(2);
+    return `\$${priceStr}`;
+};
+
+export { formatPrice };
